Extract editor blocks inline editing bridge
