+++ scripts/test-system-release-package.sh
@@ -403,6 +403,11 @@ if ! grep -qx "press-system-${version}/assets/js/editor-blocks-inline-dom-sessio
   exit 1
 fi
 
+if ! grep -qx "press-system-${version}/assets/js/editor-blocks-inline-editing-bridge.js" "${entries_file}"; then
+  echo "expected package to include editor blocks inline editing bridge code" >&2
+  exit 1
+fi
+
 if ! grep -qx "press-system-${version}/assets/js/editor-blocks-caret-session.js" "${entries_file}"; then
   echo "expected package to include editor blocks caret session code" >&2
   exit 1
